requirements.txt
discord.py
discord-components